Moved ThemeProvider from screens to index.js
