+++ Client-app/src/components/Dashboard/index.js
@@ -21,7 +21,6 @@ const Dashboard = () => {
         sx={{
           padding: 4,
           paddingBottom: "120px",
-          backgroundColor: "var(--background-secondary)",
         }}
       >
         <Grid container spacing={3}>
