+++ Client-app/src/screens/Symptoms/index.js
@@ -7,29 +7,15 @@ import Box from "@mui/material/Box";
 import SymptomsTable from "../../components/SymptomsTable";
 import BottomNav from "../../components/BottomNav";
 import MenuAppBar from "../../components/Navbar";
-import { createTheme } from "@material-ui/core/styles";
-import { ThemeProvider } from "@mui/material/styles";
-import CssBaseline from "@mui/material/CssBaseline";
-
-const theme = createTheme({
-  palette: {
-    background: {
-      default: "var(--background-secondary)",
-    },
-  },
-});
 
 const Symptoms = () => {
   return (
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
       <Box sx={{ flexGrow: 1 }}>
         <MenuAppBar />
         <SymptomsTable />
         <BottomNav />
       </Box>
-    </ThemeProvider>
   );
-};
+}
 
 export default Symptoms;
